🎨 Loudinary 함수 utils 폴더로 분리
